requirements.txt
streamlit
pymupdf
scikit-learn
pandas
numpy
opencv-python-headless
plotly
Pillow
